+++ pkg/controllers/membercluster/v1alpha1/membercluster_controller_test.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package v1alpha1membercluster
+package v1alpha1
 
 import (
 	"context"
