+++ pkg/controllers/membercluster/v1beta1/membercluster_controller_integration_test.go
@@ -2,7 +2,7 @@
 Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
-package v1beta1membercluster
+package v1beta1
 
 import (
 	"context"
