+++ pkg/controllers/membercluster/v1beta1/membercluster_controller_test.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package v1beta1membercluster
+package v1beta1
 
 import (
 	"context"
